Organize hardware integration files
