+++ src/AgIsoStackPlusPlus/include/isobus/hardware_integration/vector_asc_logger.hpp
