+++ src/AgIsoStackPlusPlus/include/isobus/hardware_integration/virtual_can_plugin.hpp
