+++ src/AgIsoStackPlusPlus/lib/Windows/InnoMakerUsb2CanLib32.lib
