+++ src/AgIsoStackPlusPlus/lib/Windows/canal32.lib
