+++ src/AgIsoStackPlusPlus/lib/Windows/canal64.lib
