+++ src/AgIsoStackPlusPlus/src/Isobus/hardware_integration/vector_asc_logger.cpp
